More Fixes
-Fixed all the crafting issues.
-Fixed Glass Doors' item texture facing the opposite direction as the
other Door items.
-Started adding biome-specific stone walls, but ran into an issue where
the Mossy variant was just the regular variant with a missing item
render. Didn't want to mess around with making a whole new block class
for them this close to the alpha, but if you think you can fix them
quick I left the code in there and just removed the register method for
them so they dont show up in-game.