Pumpkins!!!!!!
-Added Pumpkin carving and crafting functionality. Carved Pumpkins are
now crafted with a normal Pumpkin in the center, a Candle below it, and
a pair of Shears in one of the 5 empty slots at the top (the location of
the Shears determines the type of Pumpkin design).
-Once placed, Simple Carved Pumpkins can be right clicked with Shears to
carve them into their respective Average variant. The same can be done
again to turn them into their Intricate variant, but trying to carve an
already Intricate Pumpkin will just cause the face to collapse, turning
it into a Smashed Pumpkin.
-Added Smashed Pumpkins and Smashed Jack o'Lanterns.
-Added Pumpkin Lighting mechanics. works the same as Candles.