Fixes and Features!
-Added Filter Pipes! They function pretty much the same as normal
Pipes, but have two slots instead of one. Any block or item can be
placed in the bottom slot, and when items are placed into the Filter
Pipe, items that match the item in the bottom slot will be placed in
any container below the Filter Pipe, while all other items will be
placed in whichever of the other 5 directions the Filter Pipe is facing.
-Filter Pipes are crafted with 1 Pipe on the left, a Bronze Plating to
the right of it, and a Hopper below the Plating.
-Pipes are now crafted with Copper Ingots instead of Steel Ingots.
-Hoppers are crafted with a Pipe in the center instead of a Crate.
-Changed Pipe and Filter Pipe textures and recipes to be made of
different metals so that you can visually tell them apart more easily.
-Fixed Leaves and Logs not being usable in alloying Rotten Flesh into
Leather.
-Added Splitter recipe for Filter Pipes and changed the Splitter recipe
for Hoppers.
-Fixed not being able to shift-click items into Alloy Furnaces.
-Fixed shift-clicking issues with Stonecutters.
-Shift-clicking Dyes, any type of Moss, or Crushed Ice into a
Stonecutter now places it in the add-on slot.
-Virtian Soil, Barren Virtian Soil, Ashen Virtian Soil, and Virtian
Grass are now shiftable.
-Fixed Redstone Comparators not being cartable with any type of Quartz.
-Fixed Hoppers not taking items out of the Crusher’s bonus output slot.